requirements.txt
# ============================================================================
# API & WEB SERVER  
# ============================================================================
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.32.0,<0.33.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.27.0

# ============================================================================
# RAG PIPELINE - СОВМЕСТИМЫЕ ВЕРСИИ
# ============================================================================
llama-index-core>=0.12.0,<0.15.0
llama-index-embeddings-google-genai
llama-index-vector-stores-supabase>=0.3.0

# Google Gemini API
google-generativeai>=0.5.2,<0.6.0

# Pydantic
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.4.0,<3.0.0

# Database
psycopg2-binary>=2.9.9

# ============================================================================
# DOCUMENT CONVERSION (ИСПРАВЛЕНО!)
# ============================================================================
docling>=2.55.0,<3.0.0
docling-core>=2.0.0,<3.0.0
docling-ibm-models>=3.9.1,<4.0.0
docling-parse>=4.4.0,<5.0.0
pypdfium2>=4.30.0,<5.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
Pillow>=10.0.0
easyocr>=1.7.0
hf-xet>=0.1.1

# ============================================================================
# UTILITIES
# ============================================================================
psutil>=5.9.0
pandas>=2.0.0,<3.0.0
#numpy>=1.24.0,<2.0.0